19 May 2025 - Version 3.25.6.100 (18523139)

+++ Assembly-CSharp/SDG.Unturned/BlueprintOutput.cs
@@ -6,7 +6,7 @@ public class BlueprintOutput
 {
     private CachingBcAssetRef _itemRef;
 
-    public ushort amount;
+    public int amount;
 
     public EItemOrigin origin;
 
@@ -45,7 +45,7 @@ public class BlueprintOutput
         return _itemRef.IsReferenceTo(asset);
     }
 
-    public BlueprintOutput(ushort newID, byte newAmount, EItemOrigin newOrigin)
+    public BlueprintOutput(ushort newID, int newAmount, EItemOrigin newOrigin)
     {
         amount = newAmount;
         origin = newOrigin;
